requirements.txt
Django==4.1
django_auditlog==2.1.1
django_config_models==2.3.0
django_environ==0.9.0
django_extensions==3.2.0
django_import_export==2.8.0
django_mirage_field==1.4.0
django_phonenumber_field==6.3.0
django_taggit==3.0.0
setuptools==65.3.0
